Fixed test findings
