+++ pkg/controller/authentication/secret.go
@@ -18,6 +18,8 @@ package authentication
 
 import (
 	"context"
+	"reflect"
+	"time"
 
 	operatorv1alpha1 "github.com/IBM/ibm-iam-operator/pkg/apis/operator/v1alpha1"
 	utils "github.com/IBM/ibm-iam-operator/pkg/utils"
@@ -26,6 +28,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/apimachinery/pkg/util/wait"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
@@ -127,6 +130,24 @@ func (r *ReconcileAuthentication) handleSecret(instance *operatorv1alpha1.Authen
 
 	}
 
+	reqLogger.Info("Creating or updating new secret", "Certificate.Namespace", instance.Namespace, "Secret.Name", "ibmcloud-cluster-ca-cert")
+	// create or update cluster-ca-secret
+	stop := WaitForTimeout(10 * time.Minute)
+	// get ca.crt from platform-auth-secret
+	secret, err := r.waitForSecret(instance, "platform-auth-secret", stop)
+	if err != nil {
+		reqLogger.Error(err, "Failed to get Secret platform-auth-secret")
+		return err
+	}
+
+	var caCert = secret.Data["ca.crt"]
+
+	// Create or update secret ibmcloud-cluster-ca-cert with ca.crt from platform-auth-secret
+	if err := r.createClusterCACert(instance, r.scheme, ClusterSecretName, instance.Namespace, caCert); err != nil {
+		reqLogger.Error(err, "failure creating or updating ibmcloud-cluster-ca-cert secret")
+		return err
+	}
+
 	return nil
 
 }
@@ -149,4 +170,98 @@ func generateSecretObject(instance *operatorv1alpha1.Authentication, scheme *run
 		return nil
 	}
 	return newSecret
+}
+
+func (r *ReconcileAuthentication) waitForSecret(instance *operatorv1alpha1.Authentication, name string, stopCh <-chan struct{}) (*corev1.Secret, error) {
+	reqLogger := log.WithValues("Instance.Namespace", instance.Namespace, "Instance.Name", instance.Name)
+
+	reqLogger.Info("Waiting for secret", "Certificate.Namespace", instance.Namespace, "Secret.Name", name)
+
+	s := &corev1.Secret{}
+
+	err := wait.PollImmediateUntil(2*time.Second, func() (done bool, err error) {
+		if err := r.client.Get(context.TODO(), types.NamespacedName{Name: name, Namespace: instance.Namespace}, s); err != nil {
+			return false, nil
+		}
+		return true, nil
+	}, stopCh)
+
+	return s, err
+}
+
+// create ibmcloud-cluster-ca-cert
+func (r *ReconcileAuthentication) createClusterCACert(i *operatorv1alpha1.Authentication, scheme *runtime.Scheme, secretName, ns string, caCert []byte) error {
+
+	reqLogger := log.WithValues("Instance.Namespace", i.Namespace, "Instance.Name", i.Name)
+
+	// create ibmcloud-cluster-ca-cert
+	labels := map[string]string{
+		"app":                          "auth-idp",
+		"component":                    "auth-idp",
+		"app.kubernetes.io/component":  "auth-idp",
+		"app.kubernetes.io/name":       "auth-idp",
+		"app.kubernetes.io/instance":   "auth-idp",
+		"app.kubernetes.io/managed-by": "",
+	}
+	clusterSecret := &corev1.Secret{
+		TypeMeta: metav1.TypeMeta{
+			Kind: "Secret",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      secretName,
+			Namespace: ns,
+			Labels:    labels,
+		},
+		Data: map[string][]byte{
+			"ca.crt": caCert,
+		},
+	}
+
+	if err := controllerutil.SetControllerReference(i, clusterSecret, scheme); err != nil {
+		reqLogger.Error(err, "Error setting controller reference on secret:")
+	}
+	err := r.client.Create(context.TODO(), clusterSecret)
+	if err != nil {
+		if !errors.IsAlreadyExists(err) {
+			reqLogger.Error(err, "failure creating secret for ibmcloud-cluster-ca-cert")
+			return err
+		}
+
+		reqLogger.Info("Trying to update secret: as it already existed.", "Certificate.Namespace", i.Namespace, "Secret.Name", secretName)
+
+		// Update config
+		current := &corev1.Secret{}
+		err := r.client.Get(context.TODO(), types.NamespacedName{Name: secretName, Namespace: i.Namespace}, current)
+		if err != nil {
+			reqLogger.Error(err, "failure getting secret ibmcloud-cluster-ca-cert")
+			return err
+		}
+
+		// no data change, just return
+		if reflect.DeepEqual(clusterSecret.Data, current.Data) {
+			reqLogger.Info("No change found from the secret: skip updating current secret.", "Certificate.Namespace", i.Namespace, "Secret.Name", secretName)
+			return nil
+		}
+		reqLogger.Info("Found change from secret %s, trying to update it", "Certificate.Namespace", i.Namespace, "Secret.Name", secretName)
+		current.Data = clusterSecret.Data
+
+		// Apply the latest change to configmap
+		if err = r.client.Update(context.TODO(), current); err != nil {
+			reqLogger.Error(err, "failure updating secret ibmcloud-cluster-ca-cert")
+			return err
+		}
+	}
+	reqLogger.Info("Successfully created or updated secret", "Certificate.Namespace", i.Namespace, "Secret.Name", secretName)
+	return nil
+}
+
+// waitForTimeout returns a stop channel that closes when the specified timeout is reached
+func WaitForTimeout(timeout time.Duration) <-chan struct{} {
+	stopChWithTimeout := make(chan struct{})
+	go func() {
+		for range time.After(timeout) {
+		}
+		close(stopChWithTimeout)
+	}()
+	return stopChWithTimeout
 }
